[handlers] Split dose calc handlers into modules
