+++ tests/test_handlers_prompts.py
@@ -10,6 +10,8 @@ import pytest
 os.environ.setdefault("OPENAI_API_KEY", "test")
 os.environ.setdefault("OPENAI_ASSISTANT_ID", "asst_test")
 import services.api.app.diabetes.utils.openai_utils as openai_utils  # noqa: F401
+import services.api.app.diabetes.handlers.photo_handlers as photo_handlers
+import services.api.app.diabetes.handlers.sugar_handlers as sugar_handlers
 from services.api.app.diabetes.handlers import dose_calc
 from services.api.app.diabetes.utils.ui import sugar_keyboard
 
@@ -28,7 +30,7 @@ class DummyMessage:
 async def test_prompt_photo_sends_message() -> None:
     message = DummyMessage()
     update = cast(Update, SimpleNamespace(message=message))
-    await dose_calc.prompt_photo(
+    await photo_handlers.prompt_photo(
         update,
         cast(
             CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
@@ -48,7 +50,7 @@ async def test_prompt_sugar_sends_message() -> None:
         CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
         SimpleNamespace(user_data={}),
     )
-    await dose_calc.prompt_sugar(update, context)
+    await sugar_handlers.prompt_sugar(update, context)
     assert any("сахар" in t.lower() for t in message.texts)
     assert message.kwargs and message.kwargs[0].get("reply_markup") is sugar_keyboard
