+++ tests/test_register_handlers.py
@@ -25,6 +25,9 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
         dose_calc,
         profile as profile_handlers,
         reporting_handlers,
+        photo_handlers,
+        sugar_handlers,
+        gpt_handlers,
     )
 
     app = ApplicationBuilder().token("TESTTOKEN").build()
@@ -33,19 +36,19 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
     handlers = app.handlers[0]
     callbacks = [getattr(h, "callback", None) for h in handlers]
 
-    assert dose_calc.freeform_handler in callbacks
-    assert dose_calc.photo_handler in callbacks
-    assert dose_calc.doc_handler in callbacks
-    assert dose_calc.prompt_photo in callbacks
+    assert gpt_handlers.freeform_handler in callbacks
+    assert photo_handlers.photo_handler in callbacks
+    assert photo_handlers.doc_handler in callbacks
+    assert photo_handlers.prompt_photo in callbacks
     assert dose_calc.dose_cancel in callbacks
-    assert dose_calc.prompt_sugar not in callbacks
+    assert sugar_handlers.prompt_sugar not in callbacks
     assert callback_router in callbacks
     assert reporting_handlers.report_period_callback in callbacks
     assert profile_handlers.profile_view in callbacks
     assert profile_handlers.profile_back in callbacks
     assert reporting_handlers.report_request in callbacks
     assert reporting_handlers.history_view in callbacks
-    assert dose_calc.chat_with_gpt in callbacks
+    assert gpt_handlers.chat_with_gpt in callbacks
     assert security_handlers.hypo_alert_faq in callbacks
     # Reminder handlers should be registered
     assert any(
@@ -80,7 +83,7 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
 
     conv_handlers = [h for h in handlers if isinstance(h, ConversationHandler)]
     assert dose_calc.dose_conv in conv_handlers
-    assert dose_calc.sugar_conv in conv_handlers
+    assert sugar_handlers.sugar_conv in conv_handlers
     assert profile_handlers.profile_conv in conv_handlers
     assert onb_conv[0] in conv_handlers
     conv_cmds = [
@@ -91,7 +94,7 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
     assert conv_cmds and "dose" in conv_cmds[0].commands
     sugar_conv_cmds = [
         ep
-        for ep in dose_calc.sugar_conv.entry_points
+        for ep in sugar_handlers.sugar_conv.entry_points
         if isinstance(ep, CommandHandler)
     ]
     assert sugar_conv_cmds and "sugar" in sugar_conv_cmds[0].commands
@@ -114,21 +117,21 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
     text_handlers = [
         h
         for h in handlers
-        if isinstance(h, MessageHandler) and h.callback is dose_calc.freeform_handler
+        if isinstance(h, MessageHandler) and h.callback is gpt_handlers.freeform_handler
     ]
     assert text_handlers
 
-    photo_handlers = [
+    photo_handler_msgs = [
         h
         for h in handlers
-        if isinstance(h, MessageHandler) and h.callback is dose_calc.photo_handler
+        if isinstance(h, MessageHandler) and h.callback is photo_handlers.photo_handler
     ]
-    assert photo_handlers
+    assert photo_handler_msgs
 
     doc_handlers = [
         h
         for h in handlers
-        if isinstance(h, MessageHandler) and h.callback is dose_calc.doc_handler
+        if isinstance(h, MessageHandler) and h.callback is photo_handlers.doc_handler
     ]
     assert doc_handlers
 
@@ -136,14 +139,14 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
     photo_prompt_handlers = [
         h
         for h in handlers
-        if isinstance(h, MessageHandler) and h.callback is dose_calc.photo_prompt
+        if isinstance(h, MessageHandler) and h.callback is photo_handlers.photo_prompt
     ]
     assert photo_prompt_handlers
 
     sugar_cmd = [
         h
         for h in handlers
-        if isinstance(h, CommandHandler) and h.callback is dose_calc.sugar_start
+        if isinstance(h, CommandHandler) and h.callback is sugar_handlers.sugar_start
     ]
     assert not sugar_cmd
 
@@ -178,7 +181,7 @@ def test_register_handlers_attaches_expected_handlers(monkeypatch: pytest.Monkey
     gpt_cmd = [
         h
         for h in handlers
-        if isinstance(h, CommandHandler) and h.callback is dose_calc.chat_with_gpt
+        if isinstance(h, CommandHandler) and h.callback is gpt_handlers.chat_with_gpt
     ]
     assert gpt_cmd and "gpt" in gpt_cmd[0].commands
