+++ tests/test_sugar_handlers.py
@@ -8,7 +8,7 @@ from sqlalchemy import create_engine, select
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-import services.api.app.diabetes.handlers.dose_calc as dose_calc
+import services.api.app.diabetes.handlers.sugar_handlers as sugar_handlers
 from services.api.app.diabetes.services.db import Base, Entry
 
 
@@ -24,18 +24,6 @@ class DummyMessage:
 
 
 @pytest.mark.asyncio
-async def test_photo_prompt_sends_message() -> None:
-    message = DummyMessage()
-    update = cast(Update, SimpleNamespace(message=message))
-    context = cast(
-        CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
-        SimpleNamespace(),
-    )
-    await dose_calc.photo_prompt(update, context)
-    assert any("фото" in r.lower() for r in message.replies)
-
-
-@pytest.mark.asyncio
 async def test_sugar_start_initializes_pending_entry() -> None:
     message = DummyMessage()
     update = cast(
@@ -45,8 +33,8 @@ async def test_sugar_start_initializes_pending_entry() -> None:
         CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
         SimpleNamespace(user_data={}, chat_data={}),
     )
-    state = await dose_calc.sugar_start(update, context)
-    assert state == dose_calc.SUGAR_VAL
+    state = await sugar_handlers.sugar_start(update, context)
+    assert state == sugar_handlers.SUGAR_VAL
     assert context.user_data is not None
     assert context.user_data["pending_entry"]["telegram_id"] == 1
     assert context.chat_data is not None
@@ -63,12 +51,12 @@ async def test_sugar_val_non_numeric(monkeypatch: pytest.MonkeyPatch) -> None:
         CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
         SimpleNamespace(user_data={}, chat_data={}),
     )
-    await dose_calc.sugar_start(start_update, context)
+    await sugar_handlers.sugar_start(start_update, context)
 
     def fail_session() -> Any:
         raise AssertionError("DB should not be used")
 
-    monkeypatch.setattr(dose_calc, "SessionLocal", fail_session)
+    monkeypatch.setattr(sugar_handlers, "SessionLocal", fail_session)
 
     called = False
 
@@ -76,12 +64,12 @@ async def test_sugar_val_non_numeric(monkeypatch: pytest.MonkeyPatch) -> None:
         nonlocal called
         called = True
 
-    monkeypatch.setattr(dose_calc, "check_alert", fake_check_alert)
+    monkeypatch.setattr(sugar_handlers, "check_alert", fake_check_alert)
 
     message = DummyMessage("abc")
     update = cast(Update, SimpleNamespace(message=message, effective_user=SimpleNamespace(id=1)))
-    result = await dose_calc.sugar_val(update, context)
-    assert result == dose_calc.SUGAR_VAL
+    result = await sugar_handlers.sugar_val(update, context)
+    assert result == sugar_handlers.SUGAR_VAL
     assert message.replies[-1] == "Введите сахар числом в ммоль/л."
     assert not called
 
@@ -96,12 +84,12 @@ async def test_sugar_val_negative(monkeypatch: pytest.MonkeyPatch) -> None:
         CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
         SimpleNamespace(user_data={}, chat_data={}),
     )
-    await dose_calc.sugar_start(start_update, context)
+    await sugar_handlers.sugar_start(start_update, context)
 
     def fail_session() -> Any:
         raise AssertionError("DB should not be used")
 
-    monkeypatch.setattr(dose_calc, "SessionLocal", fail_session)
+    monkeypatch.setattr(sugar_handlers, "SessionLocal", fail_session)
 
     called = False
 
@@ -109,12 +97,12 @@ async def test_sugar_val_negative(monkeypatch: pytest.MonkeyPatch) -> None:
         nonlocal called
         called = True
 
-    monkeypatch.setattr(dose_calc, "check_alert", fake_check_alert)
+    monkeypatch.setattr(sugar_handlers, "check_alert", fake_check_alert)
 
     message = DummyMessage("-1")
     update = cast(Update, SimpleNamespace(message=message, effective_user=SimpleNamespace(id=1)))
-    result = await dose_calc.sugar_val(update, context)
-    assert result == dose_calc.SUGAR_VAL
+    result = await sugar_handlers.sugar_val(update, context)
+    assert result == sugar_handlers.SUGAR_VAL
     assert message.replies[-1] == "Сахар не может быть отрицательным."
     assert not called
 
@@ -129,7 +117,7 @@ async def test_sugar_val_valid_saves_and_alerts(monkeypatch: pytest.MonkeyPatch)
         CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
         SimpleNamespace(user_data={}, chat_data={}),
     )
-    await dose_calc.sugar_start(start_update, context)
+    await sugar_handlers.sugar_start(start_update, context)
 
     engine = create_engine(
         "sqlite:///:memory:",
@@ -138,19 +126,19 @@ async def test_sugar_val_valid_saves_and_alerts(monkeypatch: pytest.MonkeyPatch)
     )
     Base.metadata.create_all(engine)
     session_factory = sessionmaker(bind=engine, autoflush=False, autocommit=False)
-    monkeypatch.setattr(dose_calc, "SessionLocal", session_factory)
+    monkeypatch.setattr(sugar_handlers, "SessionLocal", session_factory)
 
     captured: dict[str, float] = {}
 
     async def fake_check_alert(update: Update, ctx: Any, sugar: float) -> None:
         captured["value"] = sugar
 
-    monkeypatch.setattr(dose_calc, "check_alert", fake_check_alert)
+    monkeypatch.setattr(sugar_handlers, "check_alert", fake_check_alert)
 
     message = DummyMessage("5.5")
     update = cast(Update, SimpleNamespace(message=message, effective_user=SimpleNamespace(id=1)))
-    result = await dose_calc.sugar_val(update, context)
-    assert result == dose_calc.END
+    result = await sugar_handlers.sugar_val(update, context)
+    assert result == sugar_handlers.END
     assert captured["value"] == 5.5
     assert any("сохран" in r.lower() for r in message.replies)
 
@@ -160,3 +148,16 @@ async def test_sugar_val_valid_saves_and_alerts(monkeypatch: pytest.MonkeyPatch)
         entry = entries[0]
         assert entry.sugar_before == 5.5
         assert entry.telegram_id == 1
+
+
+@pytest.mark.asyncio
+async def test_sugar_val_inactive_chat_returns_end() -> None:
+    message = DummyMessage("5")
+    update = cast(Update, SimpleNamespace(message=message, effective_user=SimpleNamespace(id=1)))
+    context = cast(
+        CallbackContext[Any, dict[str, Any], dict[str, Any], dict[str, Any]],
+        SimpleNamespace(user_data={}, chat_data={}),
+    )
+    result = await sugar_handlers.sugar_val(update, context)
+    assert result == sugar_handlers.END
+    assert not message.replies
